requirements.txt
python-dotenv
#install packages
langchain
langchain-openai
langchain-together
langchain-anthropic
langchain-cohere
#langchain-google-genai 
#pillow
langchain-groq
langchain-huggingface
transformers
